Added in a timer for the test.
A Timer thread will be fired when the loadQuestions() function is
called. The timer will set the endTime parameter for the Student
object when it has finished counting down. The submitAnswer()
function has been modified to check the Student object for the
value of endTime every time it is called.

TODO: Some logic has to be added to deal with the case of the user
refreshing the page, thereby calling loadQuestions() again.